Remove Document_Parser trait.
