+++ includes/REST_API/Link_Controller.php
@@ -30,7 +30,7 @@ use DOMElement;
 use DOMNodeList;
 use Google\Web_Stories\Infrastructure\HasRequirements;
 use Google\Web_Stories\Story_Post_Type;
-use Google\Web_Stories\Traits\Document_Parser;
+use Google\Web_Stories_Dependencies\AmpProject\Dom\Document;
 use WP_Error;
 use WP_Http;
 use WP_REST_Request;
@@ -43,7 +43,6 @@ use WP_REST_Server;
  * Class Link_Controller
  */
 class Link_Controller extends REST_Controller implements HasRequirements {
-	use Document_Parser;
 
 	/**
 	 * Story_Post_Type instance.
@@ -194,14 +193,16 @@ class Link_Controller extends REST_Controller implements HasRequirements {
 			return rest_ensure_response( $response );
 		}
 
-		$xpath = $this->html_to_xpath( $html );
+		$doc = Document::fromHtml( $html );
 
-		if ( ! $xpath ) {
+		if ( ! $doc ) {
 			set_transient( $cache_key, wp_json_encode( $data ), $cache_ttl );
 			$response = $this->prepare_item_for_response( $data, $request );
 
 			return rest_ensure_response( $response );
 		}
+
+		$xpath = $doc->xpath;
 
 		// Link title.
 
@@ -368,5 +369,34 @@ class Link_Controller extends REST_Controller implements HasRequirements {
 		}
 
 		return true;
+	}
+
+	/**
+	 * Retrieve content of a given DOM node attribute.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @param DOMNodeList<DOMElement>|false $query XPath query result.
+	 * @param string                        $attribute Attribute name.
+	 *
+	 * @return string|false Attribute content on success, false otherwise.
+	 */
+	protected function get_dom_attribute_content( $query, string $attribute ) {
+		if ( ! $query instanceof DOMNodeList || 0 === $query->length ) {
+			return false;
+		}
+
+		/**
+		 * DOMElement
+		 *
+		 * @var DOMElement $node
+		 */
+		$node = $query->item( 0 );
+
+		if ( ! $node instanceof DOMElement ) {
+			return false;
+		}
+
+		return $node->getAttribute( $attribute );
 	}
 }
